new script to plot UK using Geopackage format...
So the GADM maps and data site has changed the format that is shares so new R packages have been created. As such some of the code delivered in the Feb 2025 version of the R for Biochemists 101 course delivered by the Biochemical Society doesn't work. As of 5 March 2025, one of the file formats available is Geopackage which can be opened by the R package sf.