data from Webber et al (2014)
This is data used for analysis on R for Biochemists. 
Script for analysis is here: https://rforbiochemists.blogspot.com/2015/09/downloading-and-manipulating-published.html

This data file is available through this link (https://www.mcponline.org/cms/10.1074/mcp.M113.032136/attachment/eceea9a9-6b27-4282-b818-866e4cdd901f/mmc1.zip) but it is a zip file so requires multiple steps for downloading. 

Having the data here makes it a bit easier to download and use for analysis.